Extract `btsieve` crate
The state machine for transaction monitoring can be completely
standalone.
