+++ btsieve/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "btsieve"
+version = "0.1.0"
+edition = "2021"
+description = "State machine for sieving through Bitcoin transactions."
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bitcoin = "0.27"
+tracing = "0.1"
+
+[dev-dependencies]
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
