+++ daemon/Cargo.toml
@@ -8,6 +8,7 @@ anyhow = "1"
 async-trait = "0.1.52"
 bdk = { version = "0.16", default-features = false, features = ["electrum"] }
 bdk-ext = { path = "../bdk-ext" }
+btsieve = { path = "../btsieve" }
 bytes = "1"
 cached = { version = "0.33.0", default-features = false, features = ["proc_macro"] }
 conquer-once = "0.3"
